update: namespaces and cleanup
