+++ Authorization/WebApp/Context/Abstract/IContextAccessor.cs
@@ -1,0 +1,6 @@
+﻿namespace WebApp.Context.Abstract;
+
+public interface IContextAccessor<T> where T : class
+{
+    Task<T> GetContext(HttpContext httpContext);
+}
