--- Authorization/WebApp/Context/IContextAccessor.cs
@@ -1,6 +1,0 @@
-﻿namespace WebApp.Context;
-
-public interface IContextAccessor<T> where T : class
-{
-  Task<T> GetContext(HttpContext httpContext);
-}
