+++ Authorization/WebApp/Controllers/LoginController.cs
@@ -10,7 +10,8 @@ using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Mvc;
 using WebApp.Attributes;
 using WebApp.Constants;
-using WebApp.Context;
+using WebApp.Context.Abstract;
+using WebApp.Context.AuthorizeContext;
 using WebApp.Contracts.PostLogin;
 using WebApp.Controllers.Abstracts;
 using WebApp.Extensions;
